feat: Changed service and repository layers to user int instead int64

+++ app/repositories/userRepository.go
@@ -4,7 +4,7 @@ import (
 	"reby/app/models"
 )
 
-func GetUser(idUser int64) models.User {
+func GetUser(idUser int) models.User {
 	user := models.User{}
 
 	Database.First(&user, idUser)
